Uint8を極力そのまま使わないようにした。

+++ sorakado/ai/master/render_info.cc
@@ -247,7 +247,7 @@ namespace sorakado::ai::master {
                 }
                 c.a = 0xff;
             }
-            SDL_Surface *text = TTF_RenderText_Blended(font->font(), data.content.data.data(), data.content.data.length(), {static_cast<Uint8>(c.r), static_cast<Uint8>(c.g), static_cast<Uint8>(c.b), static_cast<Uint8>(c.a)});
+            SDL_Surface *text = TTF_RenderText_Blended(font->font(), data.content.data.data(), data.content.data.length(), {c.r, c.g, c.b, c.a});
             SDL_Rect r = {data.position.x * scale_ / 100, (data.position.y - scroll_) * scale_ / 100, text->w * scale_ / 100, text->h * scale_ / 100};
             SDL_BlitSurface(text, nullptr, dst->surface(), &r);
             SDL_DestroySurface(text);
